use pytorch version when defined
